+++ CHANGELOG.md
@@ -13,9 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- updated template to https://github.com/robert-koch-institut/mex-template/commit/dd987e
+
 ### Deprecated
 
 ### Removed
+
+- BREAKING: removed basic-auth, only API keys and ldap auth remain
 
 ### Fixed